# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger MMS",
    "description": "locpro's magic",
    "termsOfService": "",
    "contact": {
      "email": "nhoxloctran125@gmail.com"
    },
    "license": {
      "name": "No license",
      "url": ""
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "localhost"
    },
    {
      "url": "https://capstone-api-mms.herokuapp.com/",
      "description": "Heroku app"
    }
  ],
...
  "paths": {
...
    "/purchase-orders/cancel/{_id}": {
      "get": {
        "tags": [
          "purchaseOrders"
        ],
        "summary": "Cancel purchase order by _id",
        "operationId": "cancelPurchaseOrder",
        "parameters": [
          {
            "in": "path",
            "name": "_id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "606158fa52045c5588eef074"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Platform-Token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}